Adjust PlayerSphericalMoveemnt to be non-specific
PlayerSphericalMovement can be applied to any of the player
objects in the game -- the main logic is about moving about on
a sphere, on the grid, not hitting walls.

Wow, Unity is amazing :)

So, this commit strips any 'Player' (aka Puckmonster) logic out, which
is mostly the collision detection and pill count, and makes it a bit
less player-focussed -- the tag that defines the start position for
example.