Use a Vector2 to store player grid location
Technically this does mean we are storing as a float, but generally
since we're working with +1, -1 it's ok, and does make a lot of future
calculations (like distance between two points) much easier.

Really we should create a GridRef class, but hey ho. This is a step in the
right direction.